requirements.txt
fastapi==0.115.6
passlib==1.7.4
pydantic==2.10.3
pydantic_settings==2.6.1
PyJWT==2.10.1
sqlmodel==0.0.22
starlette==0.41.3
bcrypt==4.0.0